Move HumanGateRequestDto + converter to src/workflow/types.ts
Closes the last runtime layer violation surfaced while extracting
WorkflowManager: the manager was still importing toHumanGateRequestDto
from src/web-ui/web-ui-types.js (a value import from web-ui into
workflow). The DTO and its converter belong next to the domain
HumanGateRequest type they shadow.

- Co-locate the DTO + converter in src/workflow/types.ts.
- Repoint the two consumers (workflow-manager.ts, web-ui dispatch) at
  the new home.
- web-event-bus.ts type-imports the DTO directly from workflow/types
  (web-ui legitimately consumes workflow types).
- web-ui-types.ts no longer defines or re-exports either symbol.

After this, workflow/ → web-ui/ has only `import type { WebEventMap }`
remaining, which is correct: WebEventMap is web-flavored and stays in
web-ui. All 4283 + 338 tests pass.
