+++ src/web-ui/dispatch/workflow-dispatch.ts
@@ -25,10 +25,15 @@ import {
   type PastRunPhase,
   RpcError,
   MethodNotFoundError,
-  toHumanGateRequestDto,
 } from '../web-ui-types.js';
 import type { PastRunLoadSuccess, WorkflowManager } from '../../workflow/workflow-manager.js';
-import type { WorkflowId, WorkflowStatus, WorkflowCheckpoint, WorkflowDefinition } from '../../workflow/types.js';
+import {
+  toHumanGateRequestDto,
+  type WorkflowId,
+  type WorkflowStatus,
+  type WorkflowCheckpoint,
+  type WorkflowDefinition,
+} from '../../workflow/types.js';
 import type { WorkflowDetail } from '../../workflow/orchestrator.js';
 import { MessageLog, type MessageLogEntry, type StateTransitionEntry } from '../../workflow/message-log.js';
 import {
